Create new file
This program was developed to scrape a specific website searching for a commodity, using the Selenium web driver in Python to search through the website. The program can be ran at a specified time interval using a cronjob in the Mac OS. It the item is found in stock, and email and a text message can be sent to a desired email address, and phone number. Twilio is used for the text message functionality.